requirements.txt
fastapi
uvicorn
gemini
python-dotenv
streamlit

